create_java model file
